Add demo mode and ETW trace
Added ETW tracing for key presses (a key logger!) so it can be used
in a variety of ways, e.g. when doing performance profiling, you can
add a message of sorts via key presses if you turn on the provider.

The ETW tracing code depends on CLR 4.5, this might force some folks
to upgrade.

Added PSReadline.EnableDemoMode which uses the ETW trace to display
key presses while you type.  This is useful for talks and producing
screen captures to show the editing experience.