Speed up pasting from console
Pasting from the console was slow for 2 primary reasons:

* StringBuilder is pathologically slow if you "Insert" at the end
* Rendering requires parsing, this was happening for every character

The first issue is trival to fix - call Append when truly appending.

To address the second issue, I'm queueing up characters when we
want to read a character.  If enough characters are queued up, we
ignore requests to render if we've rendered recently.  This is
fine because we know we'll render soon because queued characters
will come along that will force us to render again.