Rework installation to go to Common Files.
Also added logic to add the custom module location to the PSModulePath.
On uninstall, the path entry is removed iff the module location ends up as
being empty after we remove our module. This makes us play nice with other
packages that want to also install in this Common Files location.