Support some accidental shift-key to act like key
Apparently people accidently hit Shift+Backspace often enough that it's annoying
when it doesn't delete the previous character.

To fix this - we'll pretend shift wasn't pressed if we see unbound keys where
the character is a control character, but the only modifier was shift.  Realistically
this is just Backspace because Shift+Return is bound, but maybe there's another
I'm missing.