Add Windows mode bindings for character searching
F3/Shift+F3 are now bound to character searching in Windows mode.
This roughly matches Visual Studio bindings.

Note that traditionally F3 copies text from previous line starting
using the position of the cursor in the current line.  It's not
hard to implement, but I'm not sure it's a useful feature, especially
with YankLastArg (which I could see mapping to F3 but probably won't.)