Replace slow DateTime.Now calls with Stopwatch (QueryPerformanceCounter)
It's an order of magnitude faster. We don't need all the fancy time info
given by DateTime.Now anyway; we just need an accurate timestamp.

There's also a correctness issue: when dealing with TimeSpans (what you
get when subtracting DateTimes), the Milliseconds property is not the
same as the TotalMilliseconds property.