Fix #99 - completion in remote sessions
When powershell has entered a remote session via Enter-PSSession,
completion was completing against the local session.

The fix is to use the remote runspace when calling the completion
api.  The remote runspace isn't exactly exposed in a clean way
from the host, so the fix is mixed between a bit of PowerShell
and C#.