Change eol behavior for NextWord/ForwardWord
Cmd and bash place the cursor at the end of the last word
instead of one past the end of the word when moving by word.

PSReadline emulated that behavior, but it's inconsistent with
word movement on the rest of the line and it's inconvenient
because you typically need to use the right arrow or end
to edit at the end of the line.

I've removed the emulated behavior, the cursor will now move
to the end of the line.