Reset current history index when done w/ recall
After a series of UpArrow/DownArrow (or whatever keys
are bound to the history search/recall), I am reseting
the current history index.  This avoids a problem where
the current line might not match the line saved in history
for that history index.

There is a small change in behavior that some might notice.
Say you are searching for some command, but you can't think
of something unique to search for in that exact command, but
you do remember something you can search for in a command entered
near that command actual command you want.

Before this change, you could use Ctrl+R to find the command near
the one you want, then use up/down arrows to get to the exact
command.

This will no longer work after this fix.  Any other fix will
likely be more complicated, so if folks ever noticed the previous
behavior and complain that it no longer works.

Fixes #97