Support menu completion
If completion is ambiguous, we will support interactively selecting
the completion using tabs and/or arrow keys after displaying the
possible completions.

This menu style completion can be cancelled with Ctrl+G or Escape.

Menu completion should be optional as not everybody will like it,
I will add an option and/or a function that makes this something
folks can customize.

Also fixed:
    * Tooltips that contain newlines now look much better
    * Undo groups handled a little better (maybe a non-issue before
      menu completion.)