Add timeout in ReadKey for event processing
The pipeline thread is blocked when PSReadline is running.
When the pipeline thread is blocked, PowerShell won't process
any events.

I added a timeout of 300ms and check for event subscribers,
if there are any, then run a bit of useless PowerShell so that
the PowerShell engine will get a chance to process events.

Fixes #133.