ValidateAndAcceptLine improvements
Script block arguments are now only validated as arguments to specific
cmdlets (which are in a user modifiable whitelist.)

A custom validation handler must now throw an exception to report an error,
returning an object is now ignored.

A sample custom validation handler was added to demonstrate typo correction.

Fixes #162