Support Cancel/Abort DigitArgument
There was an exception with input like:

a Alt+1 Ctrl+C

CancelLine was clearing the buffer, but didn't reset _current.
That fixes the exception, but it still clears the input line
when called to terminate DigitArgument, so instead we do something
similar to how interactive history search is terminated.

Fixes #184