Right click w/ multiple lines pasted as a one line
This problem only happened in ValidateAndAcceptLine
It also only happened if validation failed for any reason.

The fix is to insert a newline even if validation failed if we still have pending input.
If there is no pending input, we don't want to insert the newline, as this is the common
case and people will not want to delete the blank line from the inserted newline.