Better error for history file access issues
When there is an IO error trying to create the directory for the history file, or some other sort of error trying to read or write the history file, we shouldn't tell the user to report a bug.

Instead, the error will just report the exception message, and if it happens twice, suggest some solutions.

The error will not be reported more than twice, because that would be too annoying.
Fixes #322