Update vi mode cursor to work without Windows API
Console.CursorSize is only supported on Windows - on other platforms it
does nothing. However, when input is redirected on Windows it crashes so
we need to detect that and do something else. Ideally this would be a
user option, but at least this change supports the feature and fixes the
crash on Windows.