Fix - failing unit-tests that used an aliased member which defaults to handling upper case/lower case keys differently.
Previously, sequence of keys were interpreted using a language hack at the top of the file.
The "using _ = Keys;" statement defaulted to a class that contained both "K" and "ucK" for lower-case 'K' and upper-case 'K' respectively.
Now, "_" is a member of the class that points to a "KbdLayout" class with members "k" and "K" respectively.

Therefore, the unit-test was broken when the definition of "_" changed.