Insure to check for final token
Insure that the next token on the token stack is checked for being the
final token so that in the rare event that the EOS token is missing,
rendering doesn't attempt to access non-existant tokens.

Optimization of redundant evaluation in the subject `while` loop.

Optimization of check for last token in current state.

Fixes #911