Respect cancellation in `ReadOneOrMoreKeys()` (#3274)
Because .NET's . `Console.ReadKey()` is uncancellable, when a hosting
application cancels PSReadLine, it also has to send a key press to get
`ReadKey` to return. In this case, we want to ignore the key press (and
a user certainly would not be expecting it, as `ReadLine` has already
been canceled).