Sync ReadKeyProc thread with pipeline thread
The pipeline thread was returning before the ReadKeyProcThread had
finished processing the dummy input. This was occuring somewhat
frequently when ReadLine was invoked multiple times in a row creating a
race condition.

With these changes, the pipeline thread will wait for dummy input to be
received, dequeue the key, and continue with normal cancellation logic.