Add TerminateStragglers option (Windows)
Addresses #3745

There is a lot more detail about the problem this PR solves in the code
comments and the linked Issue. In short, this PR:

* Adds a new "TerminateStragglers" option.
* When enabled, existing console-attached PIDs are gathered into an
  "allow" list (on Windows only).
* When the shell informs us that it is time to wait for input, right
  before we disable ctrl+c signals, (if enabled) we ensure that the
  shell will not be broken by other console-attached children that may
  also be attempting to read input.

How tested: added printfs and manually triggered the problem scenario,
to verify that the code works as expected.

There are a few open questions; I will leave some comments in the PR
(for example: if someone on non-Windows enables the option, should we
write a warning, or just smile and accept it?).